Takes geth changes v1.11.14 -> v1.12.0 (#277)
* trie/

* remove rej file

* metrics/

* accounts/

* internal

* rpc/

* core/rawdb

* go.mod

* core/state

* eth/...

* core/vm core/types/ ethdb

* core/txpool

* left out

* additional core/

* should pass tests

* nits

* remove .rej file

* newTestBlockChain TODO

* Pr comments

* remove ExcessDataGas

* Add ExcessDataGas with verification

* Add verification in plugin/evm

+++ graft/coreth/eth/gasprice/feehistory.go
@@ -124,7 +124,7 @@ func (sb *slimBlock) processPercentiles(percentiles []float64) ([]*big.Int, *big
 // enforcing backend specific limitations.
 // Note: an error is only returned if retrieving the head header has failed. If there are no
 // retrievable blocks in the specified range then zero block count is returned with no error.
-func (oracle *Oracle) resolveBlockRange(ctx context.Context, lastBlock rpc.BlockNumber, blocks int) (uint64, int, error) {
+func (oracle *Oracle) resolveBlockRange(ctx context.Context, lastBlock rpc.BlockNumber, blocks uint64) (uint64, uint64, error) {
 	// Query either pending block or head header and set headBlock
 	if lastBlock == rpc.PendingBlockNumber {
 		// Pending block not supported by backend, process until latest block
@@ -149,12 +149,12 @@ func (oracle *Oracle) resolveBlockRange(ctx context.Context, lastBlock rpc.Block
 	}
 	// Ensure not trying to retrieve before genesis
 	if rpc.BlockNumber(blocks) > lastBlock+1 {
-		blocks = int(lastBlock + 1)
+		blocks = uint64(lastBlock + 1)
 	}
 	// Truncate blocks range if extending past [oracle.maxBlockHistory]
 	oldestQueriedIndex := lastBlock - rpc.BlockNumber(blocks) + 1
 	if queryDepth := lastAcceptedBlock - oldestQueriedIndex; queryDepth > maxQueryDepth {
-		overage := int(queryDepth - maxQueryDepth)
+		overage := uint64(queryDepth - maxQueryDepth)
 		blocks -= overage
 	}
 	// It is not possible that [blocks] could be <= 0 after
@@ -177,7 +177,7 @@ func (oracle *Oracle) resolveBlockRange(ctx context.Context, lastBlock rpc.Block
 //
 // Note: baseFee includes the next block after the newest of the returned range, because this
 // value can be derived from the newest block.
-func (oracle *Oracle) FeeHistory(ctx context.Context, blocks int, unresolvedLastBlock rpc.BlockNumber, rewardPercentiles []float64) (*big.Int, [][]*big.Int, []*big.Int, []float64, error) {
+func (oracle *Oracle) FeeHistory(ctx context.Context, blocks uint64, unresolvedLastBlock rpc.BlockNumber, rewardPercentiles []float64) (*big.Int, [][]*big.Int, []*big.Int, []float64, error) {
 	if blocks < 1 {
 		return common.Big0, nil, nil, nil, nil // returning with no data and no error means there are no retrievable blocks
 	}
@@ -197,7 +197,7 @@ func (oracle *Oracle) FeeHistory(ctx context.Context, blocks int, unresolvedLast
 	if err != nil || blocks == 0 {
 		return common.Big0, nil, nil, nil, err
 	}
-	oldestBlock := lastBlock + 1 - uint64(blocks)
+	oldestBlock := lastBlock + 1 - blocks
 
 	var (
 		reward       = make([][]*big.Int, blocks)
@@ -206,13 +206,13 @@ func (oracle *Oracle) FeeHistory(ctx context.Context, blocks int, unresolvedLast
 		firstMissing = blocks
 	)
 
-	for blockNumber := oldestBlock; blockNumber < oldestBlock+uint64(blocks); blockNumber++ {
+	for blockNumber := oldestBlock; blockNumber < oldestBlock+blocks; blockNumber++ {
 		// Check if the context has errored
 		if err := ctx.Err(); err != nil {
 			return common.Big0, nil, nil, nil, err
 		}
 
-		i := int(blockNumber - oldestBlock)
+		i := blockNumber - oldestBlock
 		var sb *slimBlock
 		if sbCache, ok := oracle.historyCache.Get(blockNumber); ok {
 			sb = sbCache
